requirements.txt
altair==5.1.2
asn1crypto==1.5.1
attrs==23.1.0
blinker==1.6.2
cachetools==5.3.1
certifi==2023.7.22
cffi==1.15.1
charset-normalizer==3.2.0
click==8.1.7
cloudpickle==2.0.0
colorama==0.4.6
contourpy==1.1.1
cryptography==41.0.4
cycler==0.12.1
et-xmlfile==1.1.0
Faker==19.6.2
filelock==3.12.4
fonttools==4.43.1
fpdf==1.7.2
gitdb==4.0.10
GitPython==3.1.37
idna==3.4
importlib-metadata==6.8.0
importlib-resources==6.1.0
install==1.3.5
Jinja2==3.1.2
jsonschema==4.19.1
jsonschema-specifications==2023.7.1
kaleido==0.2.1
kiwisolver==1.4.5
markdown-it-py==3.0.0
MarkupSafe==2.1.3
matplotlib==3.8.0
mdurl==0.1.2
mysql-connector-python==8.1.0
numpy==1.26.0
openpyxl==3.1.2
oscrypto==1.3.0
packaging==23.1
pandas==2.0.3
pdfkit==1.0.0
Pillow==10.0.1
platformdirs==3.8.1
plotly==5.17.0
protobuf==4.21.12
pyarrow==10.0.1
pycparser==2.21
pycryptodomex==3.19.0
pydeck==0.8.1b0
Pygments==2.16.1
PyJWT==2.8.0
pyodbc==4.0.39
pyOpenSSL==23.2.0
pyparsing==3.1.1
pypdf==3.16.2
PyPDF2==3.0.1
python-dateutil==2.8.2
pytz==2023.3.post1
PyYAML==6.0.1
referencing==0.30.2
reportlab==4.0.5
requests==2.31.0
rich==13.6.0
rpds-py==0.10.4
seaborn==0.13.0
six==1.16.0
smmap==5.0.1
snowflake-connector-python==3.2.0
snowflake-snowpark-python==1.8.0
sortedcontainers==2.4.0
streamlit==1.27.2
tenacity==8.2.3
toml==0.10.2
tomlkit==0.12.1
toolz==0.12.0
tornado==6.3.3
typing_extensions==4.8.0
tzdata==2023.3
tzlocal==5.1
urllib3==1.26.16
validators==0.22.0
watchdog==3.0.0
wkhtmltopdf==0.2
zipp==3.17.0
